Reallocate mocked license
